+++ CHANGELOG.md
@@ -2,10 +2,21 @@
 
 This CHANGELOG indicates the different versions of the `GPR` project.
 
+## Version 0.0.20
+
+- Introduce gpr_arrtpl module for templated dynamic array management
+- Introduce gpr_bin module for binary manipulation
+- Introduce gpr_buffer module for data array management
+- Rename gpr_builtin module to gpr_utils
+- Update clang-format style
+- Update Doxyfile configuration
+- Update GUIDELINE
+- Update README
+
 ## Version 0.0.19
 
 - Rename HOW-TO-DEV to GUIDELINE and rewrite it in Markdown
-- Rewrite CHANGELOG in markdown
+- Rewrite CHANGELOG in Markdown
 - Rename gpr_network module to gpr_net
 - Add optimization flag 'O3' in Release version
 - Fix Debug version